devices: move uart from hal into devices
JIRA: PD-24
