+++ devices/uart-imxrt106x/Makefile
@@ -1,0 +1,9 @@
+#
+# Makefile for uart-imxrt106x
+#
+# Copyright 2021 Phoenix Systems
+#
+# %LICENSE%
+#
+
+OBJS += $(addprefix $(PREFIX_O)devices/uart-imxrt106x/, uart.o)
