+++ devices/uart-imxrt117x/Makefile
@@ -1,0 +1,9 @@
+#
+# Makefile for uart-imxrt117x
+#
+# Copyright 2021 Phoenix Systems
+#
+# %LICENSE%
+#
+
+OBJS += $(addprefix $(PREFIX_O)devices/uart-imxrt117x/, uart.o)
