+++ devices/uart-zynq7000/Makefile
@@ -1,0 +1,9 @@
+#
+# Makefile for uart-zynq7000
+#
+# Copyright 2021 Phoenix Systems
+#
+# %LICENSE%
+#
+
+OBJS += $(addprefix $(PREFIX_O)devices/uart-zynq7000/, uart.o)
