+++ devices/uart.h
@@ -17,7 +17,7 @@
 #ifndef _SERIAL_H_
 #define _SERIAL_H_
 
-#include "types.h"
+#include "../types.h"
 
 extern void serial_init(void);
