+++ hal/armv7a/zynq7000/Makefile
@@ -21,6 +21,7 @@ GCCLIB := $(shell $(CC) $(CFLAGS) -print-libgcc-file-name)
 
 include devices/gpio-zynq7000/Makefile
 include devices/usbc-cdc/Makefile
+include devices/uart-zynq7000/Makefile
 
-OBJS += $(addprefix $(PREFIX_O)hal/$(TARGET_SUFF)/$(TARGET_SUBFAMILY)/, _startup.o hal.o zynq.o serial.o timer.o\
+OBJS += $(addprefix $(PREFIX_O)hal/$(TARGET_SUFF)/$(TARGET_SUBFAMILY)/, _startup.o hal.o zynq.o timer.o\
 		cmd.o interrupts.o)
