+++ hal/armv7m/imxrt106x/Makefile
@@ -21,6 +21,7 @@ CFLAGS+= -mfloat-abi=soft
 GCCLIB := $(shell $(CC) $(CFLAGS) -print-libgcc-file-name)
 
 include devices/usbc-cdc/Makefile
+include devices/uart-imxrt106x/Makefile
 
-OBJS += $(addprefix $(PREFIX_O)hal/$(TARGET_SUFF)/$(TARGET_SUBFAMILY)/, _startup.o hal.o imxrt.o serial.o cmd.o\
+OBJS += $(addprefix $(PREFIX_O)hal/$(TARGET_SUFF)/$(TARGET_SUBFAMILY)/, _startup.o hal.o imxrt.o cmd.o\
 		timer.o flashdrv.o rom_api.o flashcfg.o phfs-flash.o)
