+++ hal/armv7m/imxrt117x/Makefile
@@ -20,5 +20,7 @@ CFLAGS+= -mfloat-abi=soft
 
 GCCLIB := $(shell $(CC) $(CFLAGS) -print-libgcc-file-name)
 
-OBJS += $(addprefix $(PREFIX_O)hal/$(TARGET_SUFF)/$(TARGET_SUBFAMILY)/, _startup.o hal.o imxrt.o serial.o cmd.o timer.o\
+include devices/uart-imxrt117x/Makefile
+
+OBJS += $(addprefix $(PREFIX_O)hal/$(TARGET_SUFF)/$(TARGET_SUBFAMILY)/, _startup.o hal.o imxrt.o cmd.o timer.o\
 		flashdrv.o rom_api.o flashcfg.o phfs-flash.o)
