+++ plo.c
@@ -19,7 +19,7 @@
 #include "phfs.h"
 #include "timer.h"
 #include "plostd.h"
-#include "serial.h"
+#include "uart.h"
 
 #include "config.h"
